Filter non-public methods and report singleton class names properly
Non-public methods will not be included in the class map by the
inspector when running in 'implicit' mode.

When the tracer receives notice of a method call which is coming from
the singleton class of an object, report the defined_class as the normal
class of the object so that it will match up with the class map.