Properly consolidate the hierarchy
Manage parent and child relationships more tightly.
Prune after reparenting.
Perform pruning postfix and use a clone of the children array so that
the children aren't being modified in the middle of iteration.
As a cosmetic update, suppress class_name when it can be inferred from
a method's location in the feature hierarchy.