Record button loads its state dynamically
The record button loads is state (whether its recording or not) by
making an XMLHttpRequest when the page loads. This prevents HTML caching
from interfering with the correct button state display.
Added Selenium tests for the button, including page reload.
Invocations of the record button are filtered out of the AppMap.