Ensure that hooking preserves method arity
When hooking a method, we need to create a Proc that accepts the same
arguments as the original method. The only way to do this is with a
splat, i.e. `Proc.new do |*args|...`. Unfortunately, the arity of this
Proc is -1 (because it accepts a variable number of arguments), which
may or may not match the original method.

As far as I was able to determine, Ruby doesn't provide a way to get the
signature of a method and define a Proc with a matching signature. The
closest I came was to consider parsing the array returned from
`#parameters` and reconstructing the original signature. This isn't a
viable approach, because there's no way to obtain the default value for
an optional parameter.

So, instead, when hooking a method, cache its arity in the class that
owns it, and reimplement `#arity` so it uses the cached value. I opted
to do this in the extension with a hidden instance variable to hold the
cache. This avoids the possibility of interfering with any reflection
the app may be doing.