fix: Require appmap/railtie if Rails is defined
require 'appmap/railtie' will trigger autoload of 'rails/railtie', which
will allow AppMap::Railtie to work properly.

require 'appmap/railtie' when the class Rails::Railtie is first observed
does not work, because other classes such as
Rails::Railtie::Configuration are not yet available.

So, check if Rails is defined. If so, load 'appmap/railtie'. In a Rails
app, Rails is always defined when the dependency gems are loaded,
because Rails startup begins by loading 'rails/commands'.