fix: Fix compilation on macOS with Xcode 13
Xcode 13* introduces a new warning named compound-token-split-by-macro.
One of the ruby headers triggers this warning. We build our extension
with -Werror in CFLAGS, resulting in the issue described here:
https://bugs.ruby-lang.org/issues/17865 .

This doesn't appear to be a bug in the ruby header, so it seems safe to
ignore it until it's fixed. This change keeps -Werror in CFLAGS, and
just disables compound-token-split-by-macro.

* The issue indicates this behavior was actually introduced in clang 12, which
was probably part of Xcode 12.