fix: Rescue exceptions when calling Class#to_s
Calling #to_s on Rails internal classes (e.g. those used for
associations) may raise an exception. Fall back to using the class name
if that happens.

These changes also make use of a minimal Rails 7 app for testing. This
app is complete enough to show that the #to_s fix works, but not enough
to claim that we should claim Rails 7 support.