fix: Catch an uncaught EINVAL on Windows when attempting to parse source code
The `method_source` gem will happily take a source path such as
`<internal:pack>:243` and attempt to open it as a file. In this case, Windows will
raise an EINVAL error which is not caught by the gem and the error will
propagate through the agent and fail a recording.