fix: Capture HTTP requests in Rails on Rack level if possible
We used to only hook ActionController's process_action, but that
can be inaccurate if there's a middleware that intercepts the
response and modifies it, or catches an exception or an object
and does some other processing.

For example, Devise catches a throw from warden on authentication
error, then runs ActionController stack AGAIN to render a login
page, which it then modifies to change the HTTP status code.

Fixes #323