fix: Handle rack apps without engine? method in Rails
It seems in some ancient Rails versions there are routes that don't
define engine? method which we use to detect engines. Check if the
method exists before using.

Note the affected versions are all end of life; the fix is simple
but getting a test case would be more complicated so I skipped it.